# Back/docs/api/OpenAPI.yaml
openapi: 3.1.0
info:
  title: API de Gestion de Solicitudes Academicas
  version: 1.1.0
  description: >
    API REST domain-first para gestionar solicitudes academicas.
    El contrato esta alineado con los DTOs y commands de la capa de aplicacion
    del proyecto `co.edu.uniquindio.Proyecto_Avanzada`.
servers:
  - url: http://localhost:8080
    description: Entorno local
...
paths:
...
  /api/solicitudes:
    post:
      tags: [Solicitudes]
      summary: Crear solicitud
      description: >
        Crea una nueva solicitud academica asociada a un solicitante,
        indicando el tipo de solicitud, su descripcion y el canal de origen.
      operationId: crearSolicitud
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearSolicitudRequest'
      responses:
        '201':
          description: Solicitud creada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          description: Solicitud invalida o error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado o token invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operacion no autorizada para el rol actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Solicitudes]
      summary: Consultar solicitudes con filtros
      description: >
        Permite consultar solicitudes filtrando por uno o varios criterios.
        Se pueden combinar los parametros `estado`, `tipo` e
        `identificacionResponsable` en una misma peticion.
      operationId: consultarSolicitudes
      security:
        - bearerAuth: []
      parameters:
        - name: estadoSolicitud
          in: query
          required: false
          description: Estado actual de la solicitud
          schema:
            $ref: '#/components/schemas/EstadoSolicitud'
        - name: tipoSolicitud
          in: query
          required: false
          description: Tipo de solicitud
          schema:
            $ref: '#/components/schemas/TipoSolicitud'
        - name: identificacionResponsable
          in: query
          required: false
          description: Identificacion del responsable asociado
          schema:
            $ref: '#/components/schemas/Identificacion'
        - name: prioridadSolicitud
          in: query
          required: false
          description: Prioridad de la solicitud
          schema:
            $ref: '#/components/schemas/NivelPrioridad'


      responses:
        '200':
          description: Solicitudes encontradas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SolicitudResumenResponse'
        '400':
          description: Solicitud invalida o error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado o token invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operacion no autorizada para el rol actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/solicitudes/{codigo}:
    get:
      tags: [Solicitudes]
      summary: Obtener detalle de solicitud
      description: >
        Consulta la informacion completa de una solicitud a partir de su
        codigo unico, incluyendo datos generales, prioridad e historial.
      operationId: obtenerDetalleSolicitud
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitudPath'
      responses:
        '200':
          description: Detalle de la solicitud
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '401':
          description: No autenticado o token invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operacion no autorizada para el rol actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/solicitudes/{codigo}/clasificacion:
    patch:
      tags: [Solicitudes]
      summary: Clasificar solicitud
      description: >
        Clasifica una solicitud existente, definiendo su tipo y registrando
        la observacion correspondiente del usuario que realiza la accion.
      operationId: clasificarSolicitud
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitudPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClasificarSolicitudRequest'
      responses:
        '200':
          description: Solicitud clasificada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          description: Solicitud invalida o error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado o token invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operacion no autorizada para el rol actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/solicitudes/{codigo}/prioridad:
    patch:
      tags: [Solicitudes]
      summary: Priorizar solicitud
      description: >
        Asigna o actualiza el nivel de prioridad de una solicitud y guarda la
        justificacion asociada a la decision tomada.
      operationId: priorizarSolicitud
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitudPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriorizarSolicitudRequest'
      responses:
        '200':
          description: Solicitud priorizada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          description: Solicitud invalida o error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado o token invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operacion no autorizada para el rol actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/solicitudes/{codigo}/asignacion:
    patch:
      tags: [Solicitudes]
      summary: Asignar responsable a solicitud
      description: >
        Asocia un responsable a la solicitud y registra la observacion de la
        asignacion realizada por el coordinador.
      operationId: asignarResponsableSolicitud
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitudPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsignarResponsableRequest'
      responses:
        '200':
          description: Responsable asignado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          description: Solicitud invalida o error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado o token invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operacion no autorizada para el rol actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/solicitudes/{codigo}/estado:
    patch:
      tags: [Solicitudes]
      summary: Cambiar estado de solicitud
      description: >
        Actualmente esta operacion soporta la transicion hacia `ATENDIDA`,
        de acuerdo con la implementacion vigente del dominio.
      operationId: cambiarEstadoSolicitud
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitudPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CambiarEstadoRequest'
      responses:
        '200':
          description: Estado de la solicitud actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          description: Solicitud invalida o error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado o token invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operacion no autorizada para el rol actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/solicitudes/{codigo}/cierre:
    patch:
      tags: [Solicitudes]
      summary: Cerrar solicitud
      description: >
        Cierra formalmente una solicitud y almacena la observacion final
        registrada por el usuario que ejecuta el cierre.
      operationId: cerrarSolicitud
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitudPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CerrarSolicitudRequest'
      responses:
        '200':
          description: Solicitud cerrada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudDetalleResponse'
        '400':
          description: Solicitud invalida o error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado o token invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operacion no autorizada para el rol actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/solicitudes/{codigo}/historial:
    get:
      tags: [Solicitudes]
      summary: Consultar historial de una solicitud
      description: >
        Recupera la secuencia de eventos registrados sobre una solicitud,
        incluyendo estados, acciones, observaciones y responsables.
      operationId: consultarHistorialSolicitud
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CodigoSolicitudPath'
      responses:
        '200':
          description: Historial de eventos de la solicitud
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventoHistorialResponse'
        '401':
          description: No autenticado o token invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recurso no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operacion no autorizada para el rol actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CodigoSolicitudPath:
      name: codigo
      in: path
      required: true
      description: Codigo unico de la solicitud
      schema:
        type: integer
        format: int64
        examples: [1001]
...
  schemas:
    Identificacion:
      type: string
      minLength: 5
      maxLength: 20
      examples: ["1090123456"]
...
    EstadoSolicitud:
      type: string
      enum: [REGISTRADA, CLASIFICADA, EN_ATENCION, ATENDIDA, CERRADA]
      examples: [REGISTRADA, CLASIFICADA, EN_ATENCION, ATENDIDA, CERRADA]
    NivelPrioridad:
      type: string
      enum: [ALTA, MEDIA, BAJA]
      examples: [ALTA, MEDIA, BAJA]
    CanalOrigen:
      type: string
      enum: [CSU, SAC, EMAIL, PRESENCIAL, TELEFONO, PORTAL_WEB, CORREO_CERTIFICADO]
      examples: [CSU, SAC, EMAIL, PRESENCIAL, TELEFONO, PORTAL_WEB, CORREO_CERTIFICADO]
...
    TipoSolicitud:
      type: string
      enum: ["REGISTRO_ASIGNATURA","HOMOLOGACION","CANCELACION_ASIGNATURA","SOLICITUD_CUPOS","CONSULTA_ACADEMICA"]
      examples: ["REGISTRO_ASIGNATURA","HOMOLOGACION","CANCELACION_ASIGNATURA","SOLICITUD_CUPOS","CONSULTA_ACADEMICA"]
...
    CrearSolicitudRequest:
      type: object
      additionalProperties: false
      required: [tipoSolicitud, descripcion, canalOrigen, identificacionSolicitante]
      properties:
        tipoSolicitud:
          $ref: '#/components/schemas/TipoSolicitud'
        descripcion:
          type: string
          minLength: 10
          maxLength: 1000
        canalOrigen:
          $ref: '#/components/schemas/CanalOrigen'
        identificacionSolicitante:
          $ref: '#/components/schemas/Identificacion'
      examples: [{tipoSolicitud: HOMOLOGACION ,
                descripcion: "Solicito la homologacion de la asignatura Calculo Integral cursada en otra institucion.", 
                canalOrigen: PORTAL_WEB ,identificacionSolicitante: "1090123456"}]
    ClasificarSolicitudRequest:
      type: object
      additionalProperties: false
      required: [tipoSolicitud, identificacionUsuario, observacion]
      properties:
        tipoSolicitud:
          $ref: '#/components/schemas/TipoSolicitud'
        identificacionUsuario:
          $ref: '#/components/schemas/Identificacion'
        observacion:
          type: string
          minLength: 5
          maxLength: 500
      examples:
        [{tipoSolicitud: HOMOLOGACION, 
        identificacionUsuario: "80012345", 
        observacion: "La solicitud fue validada y clasificada segun el contenido adjunto."}, ]
    AsignarResponsableRequest:
      type: object
      additionalProperties: false
      required: [identificacionCoordinador, identificacionResponsable, observacion]
      properties:
        identificacionCoordinador:
          $ref: '#/components/schemas/Identificacion'
        identificacionResponsable:
          $ref: '#/components/schemas/Identificacion'
        observacion:
          type: string
          minLength: 5
          maxLength: 500
      examples:
        [{"identificacionCoordinador": "80012345",
        "identificacionResponsable": "90067890",
        "observacion": "Se asigna el caso al docente responsable del area."}]
    CambiarEstadoRequest:
      type: object
      additionalProperties: false
      required: [nuevoEstado, identificacionUsuario, observacion]
      properties:
        nuevoEstado:
          type: string
          enum: [ATENDIDA]
        identificacionUsuario:
          $ref: '#/components/schemas/Identificacion'
        observacion:
          type: string
          minLength: 5
          maxLength: 500
      examples:
        [{nuevoEstado: ATENDIDA ,
        identificacionUsuario: "90067890", 
        observacion: "La solicitud fue gestionada y se envio respuesta al estudiante."}]
    CerrarSolicitudRequest:
      type: object
      additionalProperties: false
      required: [identificacionUsuario, observacionCierre]
      properties:
        identificacionUsuario:
          $ref: '#/components/schemas/Identificacion'
        observacionCierre:
          type: string
          minLength: 5
          maxLength: 500
      examples:
        [{"identificacionUsuario": "80012345",
        "observacionCierre": "Se cierra la solicitud luego de confirmar la solucion con el estudiante."}]
    PriorizarSolicitudRequest:
      type: object
      additionalProperties: false
      required: [identificacionUsuario, nivelPrioridad, justificacion]
      properties:
        identificacionUsuario:
          $ref: '#/components/schemas/Identificacion'
        nivelPrioridad:
          $ref: '#/components/schemas/NivelPrioridad'
        justificacion:
          type: string
          minLength: 5
          maxLength: 500
      examples:
        [{identificacionUsuario: "80012345", 
        nivelPrioridad: ALTA, 
        justificacion: "El tramite debe resolverse antes del cierre de inscripciones."}]
...
    TipoSolicitudDTO:
      type: object
      additionalProperties: false
      required: [codigo, nombre]
      properties:
        codigo:
          $ref: '#/components/schemas/TipoSolicitud'
        nombre:
          type: string
          maxLength: 100
      examples:
        [{"codigo":"HOMOLOGACION","nombre":"Homologacion"}]
...
    SolicitudResumenResponse:
      type: object
      additionalProperties: false
      required: [codigo, tipoSolicitud, canalOrigen, estado, fechaHoraRegistro]
      properties:
        codigo:
          type: integer
          format: int64
        tipoSolicitud:
          $ref: '#/components/schemas/TipoSolicitudDTO'
        canalOrigen:
          $ref: '#/components/schemas/CanalOrigen'
        estado:
          $ref: '#/components/schemas/EstadoSolicitud'
        fechaHoraRegistro:
          type: string
          format: date-time
        usuarioSolicitante:
          $ref: '#/components/schemas/UsuarioResumenDTO'
        prioridad:
          oneOf:
            - $ref: '#/components/schemas/PrioridadDTO'
            - type: 'null'
      examples:
        [{"codigo": 1001,
        "tipoSolicitud": {"codigo":"HOMOLOGACION",
                          "nombre":"Homologacion"},
        "canalOrigen": "PORTAL_WEB",
        "estado": "EN_ATENCION",
        "fechaHoraRegistro": "2026-04-06T10:15:30Z",
        "usuarioSolicitante": {"nombre":"Laura Fernanda Gomez",
                              "identificacion":"1090123456",
                              "correo":"laura.gomez@uqvirtual.edu.co",
                              "activo":true,"rol":"ESTUDIANTE"},
        "prioridad":
          {"nivel":"ALTA",
          "descripcion":"Prioridad alta por cercania al cierre academico."}}]
    SolicitudDetalleResponse:
      type: object
      additionalProperties: false
      required:
        - codigo
        - tipoSolicitud
        - descripcion
        - canalOrigen
        - fechaHoraRegistro
        - estado
        - usuarioSolicitante
        - historial
      properties:
        codigo:
          type: integer
          format: int64
        tipoSolicitud:
          $ref: '#/components/schemas/TipoSolicitudDTO'
        descripcion:
          type: string
          maxLength: 1000
        canalOrigen:
          $ref: '#/components/schemas/CanalOrigen'
        fechaHoraRegistro:
          type: string
          format: date-time
        fechaCierre:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        estado:
          $ref: '#/components/schemas/EstadoSolicitud'
        usuarioSolicitante:
          $ref: '#/components/schemas/UsuarioResumenDTO'
        prioridad:
          oneOf:
            - $ref: '#/components/schemas/PrioridadDTO'
            - type: 'null'
        historial:
          type: array
          items:
            $ref: '#/components/schemas/EventoHistorialResponse'
      examples:
        [{"codigo":1001,"tipoSolicitud":
          {"codigo":"HOMOLOGACION",
          "nombre":"Homologacion"},
          "descripcion": "Solicito la homologacion de la asignatura Calculo Integral cursada en otra institucion.", 
          "canalOrigen": "PORTAL_WEB",
          "fechaHoraRegistro": "2026-04-06T10:15:30Z", 
          "fechaCierre": null,
          "estado": "EN_ATENCION",
          "usuarioSolicitante": {"nombre": "Laura Fernanda Gomez",
                                "identificacion":"1090123456",
                                "correo":"laura.gomez@uqvirtual.edu.co",
                                "activo":true,
                                "rol":"ESTUDIANTE"},
          "prioridad": {"nivel":"ALTA",
                        "descripcion":"Prioridad alta por cercania al cierre academico."},
          "historial": [{"fechaHora":"2026-04-06T10:15:30Z",
                        "estado":"REGISTRADA",
                        "accion":"CREACION",
                        "observacion":"Solicitud creada por el estudiante desde el portal web."
                        ,"responsable":{"nombre":"Laura Fernanda Gomez",
                                        "identificacion":"1090123456",
                                        "correo":"laura.gomez@uqvirtual.edu.co",
                                        "activo":true,
                                        "rol":"ESTUDIANTE"}},{"fechaHora":"2026-04-06T11:00:00Z",
                                        "estado":"CLASIFICADA",
                                        "accion":"CLASIFICADA",
                                        "observacion":"Solicitud clasificada por coordinacion.",
                                        "responsable":{"nombre":"Carlos Andres Ruiz",
                                        "identificacion":"80012345",
                                        "correo":"carlos.ruiz@uniquindio.edu.co","activo":true,"rol":"COORDINADOR"}}]}]
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
          maxLength: 100
        message:
          type: string
          maxLength: 500
        path:
          type: string
          maxLength: 300
        validationErrors:
          oneOf:
            - type: object
              additionalProperties:
                type: string
            - type: 'null'
      examples:
        [{"timestamp": "2026-04-06T11:25:00Z",
        "status": 400,
        "error": "Bad Request",
        "message": "La solicitud contiene datos invalidos.",
        "path": "/api/solicitudes",
        "validationErrors":
          {"descripcion":"La descripcion debe tener al menos 10 caracteres."}}]